requirements.txt
Django>=4.0
psycopg2>=2.9
dnspython
pymongo>=3.11
cryptography==39.0.0
djangorestframework==3.14.0
django-oauth-toolkit==1.7.0